refactor: EditProfileScreen 구조 변경 및 glueStack 적용
